+++ src/main/java/com/esri/core/geometry/OperatorExportToWkbCursor.java
@@ -36,7 +36,7 @@ public class OperatorExportToWkbCursor extends ByteBufferCursor {
 	}
 
 	@Override
-	public long getByteBufferID() {
+	public int getByteBufferID() {
 		return m_geometryCursor.getGeometryID();
 	}
